requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
transformers==5.2.0
torch==2.10.0
langgraph
